ovn-tester: Modularize test definitions.
This change encapsulates tests into their own files. Each yaml section
in the configuration file corresponds to a module in the tests/
directory. This has the following advantages:

* The logic of individual tests can be contained within its own
  module/class.
* The ovn_tester.py file can be made smaller, allowing for it to be the
  orchestrator of tests, rather than holding test logic.
* New tests can be defined without needing to modify ovn_tester.py. Just
  add the new module to tests/ and as long as the configuration is
  present, the new test will be run.
* This allows for customized ordering of test scenarios.
* This allows for test scenarios to be skipped if desired.

In this change, the cluster bringup test is handled in a special way. It
is run directly in ovn_tester.py since it must be run as the first test.
This could be made a bit less awkward by changing the config file
schema, but I sought to make this change not break things for people if
possible.

Signed-off-by: Mark Michelson <mmichels@redhat.com>