ovn-tester: Add stats collection for asynchronous contexts.
Using the typical iteration method for contexts with asynchronous code
doesn't work. Different iterations can overlap, so you can't just keep
state of a single "current" iteration.

To fix this, we need to be able to keep up with multiple iterations at
once within a context. Therefore iterations have been made a first class
object and are kept in a dictionary in the Context. A worker can
determine the current iteration by calling the new
get_current_iteration() function.

Signed-off-by: Mark Michelson <mmichels@redhat.com>