ovn-tester: Get ext_cmds working with asyncio.
The current_ext_cmds logic has some flaws with asynchronous code:
* Since all commands are run at the beginning of the specified
  iteration, and since iterations can be started in a batch, it means
  that commands might run in quicker succession than anticipated.
* Iterations typically start in the expected order, but sometimes you
  may see iterations run out of order. This could lead to strangeness if
  you assume that iterations always run in order.
* Running background commands by appending '&' to the end of the command
  doesn't work as expected with asyncssh. Control is not returned
  immediately, and so the code blocks forever.

Presumably, the only reason why it's important to start a process in the
background explicitly is so that the process can then be terminated in a
later iteration. Since asyncssh can keep track of the remote processes
it is running, it's easier to just explicitly define the number of
iterations that an external command should run for before terminating.

This commit adds a "duration" configuration option to external commands.
If the duration is 0, then the command runs immediately and we wait for
the command to complete before continuing. Otherwise, we start a remote
process. As each iteration completes, we decrease the remaining duration
until it reaches 0. At that point, the command is terminated.

This also fixes a minor bug in the ext_cmd code. The ext_cmd code was
returning the output of the command, but the only caller to execute
commands did not inspect the returned value. We no longer attempt to
return the output of the command.

Signed-off-by: Mark Michelson <mmichels@redhat.com>