do.sh: Collect data about poll intervals and database compactions.
This commit adds a couple of scripts that mines data about unreasonably
long poll intervals from logs and stores it in a human-readable format
with all the statistics including min, max, avg, 95 percentile, etc.
values per OVN daemon.  For databases it tries to exclude poll
intervals that involved compactions and also collects data about time
spent on compactions separately.  Logic for this is not ideal, but it
seems to work fine for most cases.

Statistics counted with 'datamash' utility which is a new dependency.

Example of the output:

  Unreasonably long poll intervals:

  ---------------------------------------------------------------------
      Count       Min        Max       Median      Mean   95 percentile
  ---------------------------------------------------------------------
      1073.0     1000.0     9341.0     1612.0     3483.7     8401.2
      2031.0     1011.0     9961.0     1544.0     2533.1     8157.0
      3332.0     1000.0     9501.0     1467.5     2698.3     8293.0
  ---------------------------------------------------------------------
      6436.0     1000.0     9961.0     1565.5     2777.1     8302.2

Here is a result per log file and the result when all log files
considered all together, i.e. 'Mean' on the last line is not an average
of 'Mean' values of all rows, but a 'Mean' value of the data from all
the files.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>