ovn-tester: Use logging module instead of plain prints.
'logging' module prints much more information including timestamps
which are useful while trying to match ovn-heater events with logs
from OVN daemons.

Multiline outputs from the sandbox execution are prefixed with "---\n",
because these doesn't look good when the first line is printed with all
the logging prefixes and the next ones are not.

Also, added a check that stderr is not empty to avoid printing blank
lines which are a bit annoying.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>