ovn-tester: Separate relays from other DBs.
relay containers only run a southbound database. Therefore, when we try
to enable trim on compaction for the northbound database in that
container, it results in an error. The error is benign, but it still is
better not to issue a useless command.

This commit fixes the problem by separating relay containers from other
DB containers. This way, they can get special treatment if necessary. In
this particular case, the special treatment is not to try to configure a
non-existent database.

Signed-off-by: Mark Michelson <mmichels@redhat.com>