ovn-tester: Ensure unique names for port groups and address sets.
There are two issues addressed here.

First, sub-namespaces base the name on the number of port groups created
in a given namespace. This means that if multiple namespaces are
created, then each of those namespaces attempts to create port groups
and address sets with the same names. This results in transaction
errors, meaning the port groups and address sets are not created. This
is fixed in this commit by having the port groups and address sets also
include the namespace name.

Second, if cleanup is not enabled for tests, then subsequent tests would
create namespaces with identical names. This results in the same
transaction errors. This is fixed in this commit by ensuring all network
policy tests create uniquely named namespaces.

Signed-off-by: Mark Michelson <mmichels@redhat.com>