ovn-tester: Don't use pings for testing if port is up.
Pings are used for two reasons in ovn-tester:
1) To test that a network policy is being enforced.
2) To determine that a newly-created port is up.

The first reason is a good use of pings since it ensures that the
packets are able (or unable) to make a specific roundtrip.

The second reason is useful, but it's also used to determine the time
that it takes for the port in question to go from down to up. This is
where it falls short.

First, it's vital that the ping attempts begin ASAP after provisioning
the new port. If we do not attempt to ping for a long time, then the
pings are likely to succeed immediately, and they don't reflect well on
how long it actually took for the port to be up. For synchronous code,
this isn't much of a concern, since only a single worker is ever
"active" at once. It's easy to ensure that the ping is attempted
immediately after binding the port. For asynchronous code, though, this
guarantee is harder to make. If there are enough active tasks, it may be
several seconds (or even minutes) between when the port is bound and
when we start pings.

Second, multiple pings may be necessary before we can make the
determination the port is up. Like with the problem in the above
paragraph, it's hard for asynchronous code to ensure that it can attempt
the follow-up pings in a timely manner. So even if we were to fix the
first problem and immediately start pinging after binding a port, we
also have to ensure that all follow-up pings happen ASAP.

A much better means of measuring the time it takes for a port to come up
is to use the external_ids:ovn-installed-ts value in the OVS Interface
table. This records the timestamp when the ovn-controller instance
claimed the binding. If we record when we provisioned the port, then we
can subtract this time from the ovn-installed-ts to find out the amount
of time that passed between when we inserted the port into the
northbound database and when the port was claimed by an ovn-controller.

Another nice thing is that we can perform this calculation long after
the port has been claimed and it will be the same. Therefore, we can
retrospectively find the wait times for all ports after a test has
completed, rather than performing mid-test pings.

This commit adds some new methods to WorkerNode and Cluster to wait and
measure the time it took for a port to be claimed. It then changes the
base_cluster_bringup, density_heavy, density_light, and cluster_density
tests to use this method for determining the amount of time the port
took to come up.

One thing this commit "breaks" a bit is the logging for whether an
iteration succeeded or failed. Since we wait until after the test is
completed to calculate the wait times, it means that iterations will
complete and potentially log that they were successful, but our
calculation afterward can end up failing the iteration.

One final note: with this commit, the ping_ports() members of WorkerNode
and Cluster are no longer used. They could be removed altogether, but
they could prove useful at some point, so I've kept them in the code.

Signed-off-by: Mark Michelson <mmichels@redhat.com>