ovn-tester: Allow configurable connections.
asyncio has the potential to create many tasks that each may be issuing
SSH commands. When using a single SSH connection, asyncssh has to
multiplex the commands, and it can lead to slowdown of tests.

This commit adds a configuration option for the central node that allows
for multiple SSH connections to be used instead. It also sets a default
of 30 connections to the central node. Doing this greatly increases the
speed at which tests run.

The worker nodes have a hard-coded 5 connections per node. This could
potentially be configurable too, but worker nodes tend not to have as
much traffic sent to them, so there is typically not a need to have a
lot of connections.

Signed-off-by: Mark Michelson <mmichels@redhat.com>