do.sh: Add lport binding to ovn-installed latency to mined data.
Difference per logical port between port being bound to the node
and ovn-controller marking it as ovn-installed, including usual
min, max, avg, 95%ile.  Ports that was never actually installed
are also counted.

Results are in the mined-data/binding-to-ovn-installed-latency.

The metric should be taken with a grain of salt, because ports
are bound (OVS port with iface-id created) after they were already
created in Nb DB.  And for bulk port creations, binding happens
in bulk as well afterward.  The value can be more or less treated
as an end to end OVN latency for ports created during the test
phase though.

Useful to spot issues like very high latency or ports not being
installed at all.

ovn-heater might be changed in the future to always bind ports
right after creation to make the metric more useful.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>