ovn-tester: Use ovsdbapp in place of ovs-vsctl
This adds the basic building blocks for using ovsdbapp in ovn-tester,
and converts usage of ovs-vsctl to ovsdbapp.

There is still one call to ovs-vsctl in the code, since we need to set
up the SSL record before we can connect to the database over SSL. This
could be avoided if ovn-fake-multinode were updated to do this when SSL
is enabled.