cirrus: Check low-scale run in Cirrus CI.
Add a Cirrus CI configuration YAML file with the job to run a minimal
low-scale test.  It is practically not possible or extremely hard to
do the same in GitHub Actions for few reasons:

 - In GHA we're already inside of a container.
 - This container is ubuntu-based.
 - ovn-heater dosn't support debian-based distributions so we need to
   run a nested container.
 - Using a 'container:' config we can run a fedora container, but the
   container command is always overwritten, so no way to run a
   container with systemd as init.
 - I failed to start any container with systemd as init manually.
 - 'runner' is not a root, most of the operations require 'sudo',
   but we will ssh to localhost without using sudo afterwards, even
   if the user is replaced with 'runner' in ansible configuration.
 - Many more environment quirks that doesn't allow you to do what
   you actually need.

Cirrus CI's default containers are similar to GHA, but with Cirrus CI
we're able to just start a plain fedora-cloud VM and have a full access
to it as root.

'runtime' is cached to speed up builds.  Command logging enabled to
simplify troubleshooting.  Container image is cached as well.
'runtime' is archived to avoid symlink breakage.

Since we're re-constructing and re-packing many parts of the 'runtime',
the cache will be re-uploaded every time.  But, at least, we're not
re-building it, which is the main part.

Added a check for already existing venv to allow re-using a cached
version.

Reported-at: https://github.com/dceara/ovn-heater/issues/126
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>