do.sh: Don't errexit in the post-processing stage.
If we managed to run the test and collect its logs it's good enough to
determine if the test passed or failed.  The post-processing stage
shouldn't fail but in case it does (e.g., because of a bug in the
processing tools) it should not trigger the whole test to fail.

Run the post-processing in a separate subshell with errexit disabled.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>