ovn-tester: Containerize the tester.
If the tester is going to use ovsdbapp, then the tester needs to be able
to connect directly to the other nodes, not the hosts running the node
containers. The best method for doing this is to make a container to run
the test in, and then bridge it to the other node containers.

This change accomplishes that. Instead of installing ovn-tester
dependencies in a virtual environment, we now create a container to run
the ovn-tester in. The tester container can be run anywhere, and its
configuration is handled in the physical deployment configuration.

Despite the fact that the containerized tester is intended to
communicate directly with the DBs on the test nodes, there is still a
need for the tester to ssh to the hosts. This is because we still need
to be able to do non-DB operations, such as creating network namespaces.
As such, the tester needs to be given an SSH private key that can be
used to communicate with the hosts. This is configured in the physical
deployment configuration. Perhaps a future improvement could eliminate
all SSH usage in ovn-tester, but for now it's still necessary.

One final thing to note is that prior to this change, the tester and the
stats collection scripts ran in the same virtualenv. Now the tester runs
in a container, and the virtualenv is only necessary for stats
collection afterwards. Therefore, the virtualenv's use is much more
narrow now.