Bring the base cluster up with ansible.
This change add a playbook that will:

 1. Start an ovn-tester container on the first IP in mgmt_net.
 2. Start central nodes on IPs starting with the second IP.
 3. Start worker nodes equally distributed across all worker
    hosts.  With IPs starting from the (1 + n_central + n_relay)th.
 4. Start the process monitor script in each container, including
    the ovn-tester.
 5. Configure SSL remotes on ovsdb-server processes on workers.

Playbook requires an inventory in the following format:

  [tester_hosts]
  <tester-host> ansible_user=root become=true internal_iface=lo

  [central_hosts]
  <central-host> ansible_user=root become=true internal_iface=lo

  [worker_hosts]
  <worker-host-1> ansible_user=root become=true internal_iface=lo
  ...
  <worker-host-N> ansible_user=root become=true internal_iface=lo

  [all:vars]
  ovn_fake_multinode_target_path=/root/ovn-heater/runtime
  enable_ssl=yes
  clustered_db=yes
  monitor_all=yes
  etcd_cmd=no
  mgmt_net=192.16.0.0/16
  n_relays=3
  n_central=3
  n_nodes=20
  datapath_type=system

'all' is a generic group that covers all the hosts in all groups.
Variables, of course, can be provided via cmdline instead of being
defined in the inventory.  And this is how it is done in do.sh.

IPs for worker containers were 1-off from database containers,
but we can't do that anymore, because ovn-tester has to be in sync
with the ansible playbook allocating these IPs.

ovn-tester no longer needs to call ovn-fake-multinode, set up SSL
or start process monitoring.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>