Remove timedelta stuff from latency.py
Since we deploy the tester as a pod, we can be consistent with the time
zones across all pods, and we don't need to try to factor in the local
time zone of the orchestrator.

OK, and I may have also "touched up" latency.py in places, because I
find morphing the types in a dictionary to be abhorrent.