config: Centralize default YAML.
Test scenario configuration is fairly lean since all configuration
values have some default value. In the good old days, the default values
for configuration items could be dictated solely in ovn-tester.

With the containerization of the tester, now do.sh sometimes needs to
use configuration values as well. This means that do.sh sometimes also
needs to be able to use the default values. However, do.sh has no idea
what these default values are, so they end up getting duplicated by
developers.

This change seeks to solve this problem. A new python script is now the
arbiter of default values. When running a test, do.sh will take the
test configuration, run it through this script, and generate a yaml file
that has all default values filled in. This is then passed to ovn-tester
as the configuration file. ovn-tester can now accept the configuration
file as-is and not have to try to fill in any default values.