cirrus: Fail CI if there are python stack traces in the log.
If the test script fails, it doesn't fail with an error code.
It continues to gather logs and mine data, so we have a clean
exit code in the end.

We look for iteration failures afterwards, but we don't look
for hard failures in python code.

Fix that by grepping for python 'Traceback' and failing if
it is found.

Also looking for 'Failed to run test' message in case we missed
some other type of failure.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>