install-fake-multinode.yml: Copy ovn-fake-multinode from the orchestrator.
Instead of cloning the same repo from each physical host it can
be copied from the orchestrator node, where we already have it.

Simultaneous git clone from multiple physical nodes frequently
results in GitHub connection timeouts.

'synchronize' module also takes care of updating existing copy
and removing unnecessary files, so no need to delete the directory
beforehand.  And this avoids a chicken and egg problem if the
orchestrator is also one of the nodes, because synchronization
of the directory to itself is possible, while removal + copy
is not.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>