ovn-tester: Fix exceptions while adding LBs after leadership transfer.
Python IDL reconnects on leadership transfer before receiving reply for
an in-progress transaction.  But the transaction might, and likely will,
actually succeed.  So the LB may already exist after re-connection.
Let ovsdbapp know about that, so it doesn't throw exceptions when this
happens.

This doesn't solve all the possible issues with other types of
resources, but LBs are the most common after ports that already have
a proper handling in place.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>