process-monitor: Better identify process names.
Currently it's hard to say which ovsdb process is which, we can
only guess based on the process pid.  Look deeper into the process
command line arguments and determine the name based on the name
of a pid file or a name of the python script.

For example:

  Before:
    ovsdb-server-ovn-central-az0-1-138
    ovsdb-server-ovn-central-az0-1-153
    ovsdb-server-ovn-central-az0-1-64
    ovsdb-server-ovn-central-az0-1-80
    ovn-northd-ovn-central-az0-1-91
    ovn-ic-ovn-central-az0-1-184
    python3-ovn-tester-50
    python3-ovn-tester-51

  After:
    ovn_ic_nb_db|ovn-central-az0-1|138
    ovn_ic_sb_db|ovn-central-az0-1|153
    ovnnb_db|ovn-central-az0-1|64
    ovnsb_db|ovn-central-az0-1|80
    ovn-northd|ovn-central-az0-1|91
    ovn-ic|ovn-central-az0-1|184
    process-monitor|ovn-tester|50
    ovn_tester|ovn-tester|51

Using '|' as a delimiter for easier parsing in future commits.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>