process-stats: Add support for graphs with aggregate resource usage.
We take all the stats and generate sum/mean/max/min graphs per process
type as well as an extra graph for all OVN processes together and
another one for all OVS processes together.

Resulted HTML also contains the peak of the sum of RSS and the peak of
the sum of CPU usage for all OVN processes together.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>