log-collector.sh: Fix collection of OF rules.
While dumping OF groups was limited to the last $limit containers on
each node, that wasn't the case for OF rules.  Impose the same limit
to avoid wasting disk space for mostly similar outputs from all virtual
nodes on large scale (500 node) runs.

Also, reduce the $limit from 10 to 2 and pre-archive the dumped OF
rules/groups text files for the same reason.

Fixes: 3fa30c388fb5 ("scripts: log-collector: Add ovs open-flows and group dump to log results")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>